+++ src/main/java/com/yzgeneration/evc/common/implement/port/UuidHolder.java
@@ -1,4 +1,4 @@
-package com.yzgeneration.evc.common.service.port;
+package com.yzgeneration.evc.common.implement.port;
 
 public interface UuidHolder {
     String random();
